Getting CM's natural mother's age at first birth
Computed as a variable for all natural children of that mother, along
with the the age at which the mother gave birth to each individual
child.

Finally, this file contains an intermediate section where all natural
children of CM's natural mother are extrapolated. This can be used to
manipulate twins etc,